requirements.txt
asgiref==3.11.0
Django==6.0.1
djangorestframework==3.16.1
dj-database-url==2.2.0
gunicorn==23.0.0
sqlparse==0.5.5
tzdata==2025.3
whitenoise==6.7.0
